Deep-Learning/requirements.txt
editdistance
einops
pandas
numpy
scipy
hydra-core
scikit-learn
tqdm
matplotlib
seaborn
coloredlogs
tensorboard
colorcet
opencv-python
